Fix doc variable naming in doc_mut_copy
Since introduction of `yyjson_mut_doc_mut_copy` function,
the variable name `i_doc` lost its purpose, since `i` in the name
means `immutable`, but this is false for `yyjson_mut_doc`.

Rename `i_doc` to `doc` to match function declarations in `yyjson.h`
and fix this broken logic in function naming.

It could possible be changed to `m_doc` in `yyjson_mut_doc_mut_copy`,
but this is already used for internal variable, so change it to `doc` instead.