Fix val variable naming in doc_mut_copy
The same reason as in previous commit.

First of all, change `ctn` to `val` in function name, since it can copy
any value type, not only containers.

Second, it takes `yyjson_mut_val` as argument, so it should be named
`m_vals` since it's mutable. So, rename `i_val` to `m_vals`.