Simplify containers traversal in yyjson_mut_val_mut_copy
The mutable object or array stores all sub-values in a circular linked list,
so we can traverse them in the same loop. The traversal starts from the last
item, continues with the first item in a list, and ends with the second
to last item, which needs to be linked to the last item to close the circle.